Signer: expose padUnsignedMultisigs
Introduces optional behavior to the transaction signer.
Calling `$signer->padUnsignedMultisigs(true)` will
cause the input signer to fill a partially signed
multisignature transaction scriptSig with a OP_0 byte
if the signature is missing.

If the transaction is fully signed, the scriptSig is
serialized normally.

It does NOT affect signature extraction at this time,
this will follow in a separate commit